// src/generated/api/catalog/models/attributeDataType.ts
/**
 * Generated by orval v7.13.2 🍺
 * Do not edit manually.
 * DCTD-UTC Commerce API
 * Contract for storefront and admin applications
 * OpenAPI spec version: 1.0.0
 */

export type AttributeDataType = (typeof AttributeDataType)[keyof typeof AttributeDataType];

// eslint-disable-next-line @typescript-eslint/no-redeclare
export const AttributeDataType = {
  TEXT: 'TEXT',
  NUMBER: 'NUMBER',
  BOOLEAN: 'BOOLEAN',
  OPTION: 'OPTION',
} as const;